# src/HouseholdLedger.Api/openapi/v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "HouseholdLedger.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/accounts/{accountId}/transactions": {
      "get": {
        "tags": [
          "AccountTransactions"
        ],
        "summary": "Gets one account and its matching transactions newest first.",
        "parameters": [
          {
            "name": "accountId",
            "in": "path",
            "description": "The route-selected account identifier.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "From",
            "in": "query",
            "description": "Gets or initializes the inclusive first ledger date.",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "To",
            "in": "query",
            "description": "Gets or initializes the inclusive final ledger date.",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "Classification",
            "in": "query",
            "description": "Gets or initializes the required expense classification.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "MinimumAmount",
            "in": "query",
            "description": "Gets or initializes the inclusive minimum stored amount.",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
              "type": [
                "number",
                "string"
              ],
              "format": "double"
            }
          },
          {
            "name": "MaximumAmount",
            "in": "query",
            "description": "Gets or initializes the inclusive maximum stored amount.",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
              "type": [
                "number",
                "string"
              ],
              "format": "double"
            }
          },
          {
            "name": "Search",
            "in": "query",
            "description": "Gets or initializes text matched against visible transaction fields.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountTransactionHistoryResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/income-receipts": {
      "post": {
        "tags": [
          "IncomeReceipts"
        ],
        "summary": "Confirms that income was received and records its actual destinations.",
        "requestBody": {
          "description": "The cancellation token.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmIncomeReceiptRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmIncomeReceiptRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmIncomeReceiptRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncomeReceiptResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/income-receipts/{receiptId}": {
      "put": {
        "tags": [
          "IncomeReceipts"
        ],
        "summary": "Corrects a confirmed receipt's actual date, amount and destinations.",
        "parameters": [
          {
            "name": "receiptId",
            "in": "path",
            "description": "The receipt identifier.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "The cancellation token.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateIncomeReceiptRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateIncomeReceiptRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateIncomeReceiptRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncomeReceiptResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "IncomeReceipts"
        ],
        "summary": "Deletes a mistaken confirmed receipt and its account allocations.",
        "parameters": [
          {
            "name": "receiptId",
            "in": "path",
            "description": "The receipt identifier.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "IncomeReceipts"
        ],
        "summary": "Gets a confirmed income receipt.",
        "parameters": [
          {
            "name": "receiptId",
            "in": "path",
            "description": "The receipt identifier.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncomeReceiptResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AccountResponse": {
        "required": [
          "id",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The account identifier.",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "description": "The account display name."
          }
        },
        "description": "Represents one persisted account."
      },
      "AccountTransactionHistoryResponse": {
        "required": [
          "account",
          "transactions"
        ],
        "type": "object",
        "properties": {
          "account": {
            "description": "The selected account.",
            "$ref": "#/components/schemas/AccountResponse"
          },
          "transactions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExpenseTransactionResponse"
            },
            "description": "Every transaction owned by the account in newest-first order."
          }
        },
        "description": "Represents one account and its complete transaction history."
      },
      "ConfirmIncomeReceiptRequest": {
        "required": [
          "receivedDate",
          "amount",
          "allocations"
        ],
        "type": "object",
        "properties": {
          "receivedDate": {
            "type": "string",
            "description": "The actual date the income was received.",
            "format": "date"
          },
          "amount": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "description": "The positive amount received.",
            "format": "double"
          },
          "allocations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IncomeAllocationRequest"
            },
            "description": "The actual accounts holding the receipt."
          },
          "requestId": {
            "type": [
              "null",
              "string"
            ],
            "description": "An optional client-generated retry token; reuse only for the identical confirmation.",
            "format": "uuid"
          }
        },
        "description": "Records income the user confirms was actually received."
      },
...
      "ExpenseTransactionResponse": {
        "required": [
          "id",
          "accountId",
          "accountName",
          "date",
          "amount",
          "classification"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The opaque transaction identifier.",
            "format": "uuid"
          },
          "accountId": {
            "type": "string",
            "description": "The owning account identifier.",
            "format": "uuid"
          },
          "accountName": {
            "type": "string",
            "description": "The owning account name."
          },
          "date": {
            "type": "string",
            "description": "The ledger date.",
            "format": "date"
          },
          "amount": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "description": "The positive USD amount.",
            "format": "double"
          },
          "classification": {
            "type": "string",
            "description": "The Kakeibo-inspired classification."
          },
          "description": {
            "type": [
              "null",
              "string"
            ],
            "description": "The optional normalized expense description."
          }
        },
        "description": "Describes one persisted USD expense."
      },
...
      "IncomeAllocationRequest": {
        "required": [
          "accountId",
          "amount"
        ],
        "type": "object",
        "properties": {
          "accountId": {
            "type": "string",
            "description": "The destination account identifier.",
            "format": "uuid"
          },
          "amount": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "description": "The positive amount allocated to the account.",
            "format": "double"
          }
        },
        "description": "Requests the allocation of one income receipt amount to an account."
      },
      "IncomeAllocationResponse": {
        "required": [
          "accountId",
          "amount"
        ],
        "type": "object",
        "properties": {
          "accountId": {
            "type": "string",
            "description": "The destination account identifier.",
            "format": "uuid"
          },
          "amount": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "description": "The allocated amount.",
            "format": "double"
          }
        },
        "description": "Describes one income allocation to an account."
      },
      "IncomeReceiptResponse": {
        "required": [
          "id",
          "scheduleId",
          "payDate",
          "netIncome",
          "allocations"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The receipt identifier.",
            "format": "uuid"
          },
          "scheduleId": {
            "type": [
              "null",
              "string"
            ],
            "description": "The optional expectation schedule identifier.",
            "format": "uuid"
          },
          "payDate": {
            "type": "string",
            "description": "The receipt pay date.",
            "format": "date"
          },
          "netIncome": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "description": "The received net income.",
            "format": "double"
          },
          "allocations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IncomeAllocationResponse"
            },
            "description": "The immutable account allocation snapshot."
          }
        },
        "description": "Describes a confirmed income receipt."
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "UpdateIncomeReceiptRequest": {
        "required": [
          "receivedDate",
          "amount",
          "allocations"
        ],
        "type": "object",
        "properties": {
          "receivedDate": {
            "type": "string",
            "description": "The actual past or current receipt date.",
            "format": "date"
          },
          "amount": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "description": "The positive amount received.",
            "format": "double"
          },
          "allocations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IncomeAllocationRequest"
            },
            "description": "The corrected account destinations, totaling the amount."
          }
        },
        "description": "Corrects an explicitly confirmed income receipt."
      },
...
      "ValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      }
    }
  },
...
}